Restructure unit test directory
- UT directory now looks like this:
sdk/test
--> app
       --> browser (contains browser unit test files)
       --> cordova (contains cordova unit test app)
--> misc
--> tests
       --> shared (test code shared between all targets (browser and cordova)

- Also, updated gruntfile and package.json to add a few useful scripts (buildbrowser, buildcordova, etc)
