+++ Gruntfile.js
@@ -25,8 +25,8 @@ module.exports = function(grunt) {
             ],
             // Entry points common to tests for all platforms
             testcore: [
-                'sdk/test/framework/*.js',
-                'sdk/test/*.js'
+                'sdk/test/misc/**/*.js',
+                'sdk/test/tests/shared/**/*.js'
             ],
             // List of all javascript files that we want to validate and watch
             // i.e. all javascript files except those that are installed, generated during build, third party files, etc
@@ -35,11 +35,9 @@ module.exports = function(grunt) {
                 'sdk/src/**/*.js',
                 'sdk/test/**/*.js',
                 '!**/[gG]enerated/*.js',
-                '!sdk/test/cordova/platforms/**',
+                '!sdk/test/app/cordova/platforms/**',
                 '!sdk/test/**/bin/**',
-                '!sdk/test/**/plugins/**',
-                '!**/node_modules/**',
-                '!**/MobileServices.*js'
+                '!sdk/test/**/plugins/**'
             ]
         },        
         jshint: {
@@ -100,24 +98,21 @@ module.exports = function(grunt) {
                     '<%= files.web %>',
                     '<%= files.testcore %>'
                 ],
-                dest: './sdk/test/Generated/Tests.js',
+                dest: './sdk/test/app/browser/generated/tests.js',
                 options: {
                     preBundleCB: definePlatformMappings( [ { src: '**/*.js', cwd: __dirname + '/sdk/src/Platforms/web', expose: 'Platforms' } ] )
                 }
             },
-// Uncomment this when Cordova UTs are added
-//            cordovaTest: {
-//                src: [
-//                    '<%= files.cordova %>',
-//                    './sdk/test/web/js/TestFrameworkAdapter.js',
-//                    './sdk/test/web/js/TestClientHelper.js',
-//                    '<%= files.testcore %>'
-//                ],
-//                dest: './sdk/test/cordova/www/js/Generated/Tests.js',
-//                options: {
-//                    preBundleCB: definePlatformMappings( [ { src: '**/*.js', cwd: __dirname + '/sdk/src/Platforms/web', expose: 'Platforms' } ] )
-//                }
-//            },
+            cordovaTest: {
+                src: [
+                    '<%= files.cordova %>',
+                    '<%= files.testcore %>'
+                ],
+                dest: './sdk/test/app/cordova/www/js/generated/tests.js',
+                options: {
+                    preBundleCB: definePlatformMappings( [ { src: '**/*.js', cwd: __dirname + '/sdk/src/Platforms/web', expose: 'Platforms' } ] )
+                }
+            },
         },
         copy: {
             web: {
@@ -159,7 +154,7 @@ module.exports = function(grunt) {
         
     // Default task(s).
     grunt.registerTask('build', ['concat', 'browserify', 'uglify', 'copy', 'jshint']);
-    grunt.registerTask('buildweb', ['concat', 'browserify:web', 'browserify:webTest', 'copy:web']);
+    grunt.registerTask('buildbrowser', ['concat', 'browserify:web', 'browserify:webTest', 'copy:web']);
     grunt.registerTask('buildcordova', ['concat', 'browserify:cordova', 'copy:cordova']);
 
     grunt.registerTask('default', ['build']);
