+++ sdk/test/app/browser/index.html
@@ -6,18 +6,6 @@
         <link href="css/styles.css" rel="stylesheet" />
     </head>
     <body>
-        <button id="edit-test-config">Configure</button>
-        <form id="test-config">
-            <fieldset>
-                <legend>Configuration</legend>
-                <p>
-                    <label>Runtime URI:</label>
-                    <input id="runtime-uri" name="runtime-uri" />
-                </p>
-                <button type="submit">Update</button>
-            </fieldset>
-        </form>
-
         <div id="qunit"></div>
 
         <script src="http://code.jquery.com/qunit/qunit-1.11.0.js"></script>
